Multiple serious errors in the article, some of legal impact
First, the ARM term is not only an unapproved initialism, but one for which we were successfully sued in the past, so we had to remove it. Please do not use the term ARM in our documentation. It was also referring to a different product (Azure Resource Manager), it was never approved for Azure Rights Management in the first place. The right abbreviation for Azure Rights Management is Azure RMS (and we are well aware that the S there comes out of nowhere, but that's the initialism approved by Marketing and we have to stick with it). 
Second, Azure RMS doesn't use AIP as claimed in the note, it is the other way around. 
Third, there is a mixup in the AD RMS section, OME is the product that's not compatible with AD RMS, not Azure RMS. 
In addition to that I fixed a few smaller errors and omissions. But if nothing else, the removal of ARM MUST be approved, unless we want to spend a lot of money on legal fees again.