Update advanced-spam-filtering-asf-options.md
We need to pull this article. The feature is no longer functioning right and we are not going to make the investments to fix it as there is other more modern tech to replace it already in play. As such, this article needs to go while we put in the work to remove the feature from customer view.